requirements.txt
# Reddit RSS feeds (no API needed!)
feedparser>=6.0.10
requests>=2.31.0

# Telegram client (for reading groups)
telethon>=1.36.0

# Telegram bot (for sending notifications)
python-telegram-bot>=21.0

# AI classification
anthropic>=0.39.0

# Config
pyyaml>=6.0

# Utils
python-dateutil>=2.8.0
